Fix(UI): Adjusted padding and spacing & Organized IssueGatePass files
